Refactoring, reduce duplicate code
