+++ e2e/features/layer_visibility.feature
@@ -1,21 +1,14 @@
 Feature: PermaplanT planting
     As a user I want to be able to change the layer visibilities
 
-  Background:
-    Given I am on the SUTLayerVisibility map page
-
     Scenario: Successfully change plant layer visibility
-        When I turn the plant layer visiblity off
+        Given I am on the SUTLayerVisibility map page and I have selected the plant layer
+        When I plant something
+        And I turn the plant layer visibility off
         Then all plants are invisible
 
     Scenario: Successfully change base layer visibility
-        When I turn the base layer visiblity off
+        Given I am on the SUTLayerVisibility map page and I have selected the base layer
+        When I chose a background
+        And I turn the base layer visiblity off
         Then the base layer is invisible
-
-    Scenario: Successfully change plant layer opacity
-        When I lower the opacity of the plant layer
-        Then all plants opacity gets lowered
-
-    Scenario: Successfully change base layer opacity
-        When I lower the opacity of the base layer
-        Then the base layers opacity gets lowered
