+++ e2e/features/search_plants.feature
@@ -1,18 +1,18 @@
 Feature: Searching for Plants
 
-  Background:
-    Given I am on the SUTSearching map page and I have selected the plant layer
+    Background:
+        Given I am on the SUTSearching map page and I have selected the plant layer
 
-  Scenario Outline: Searching for plants with exact/partial matches
-    When I type <plant> into the search box
-    Then the app should display <result> as first match
-    Examples:
-			| plant   | result                      |
-			| tomato  | tomato Solanum lycopersicum |
-			| tato    | potato Solanum tuberosum    |
-			| nion    | onion Allium cepa           |
-			| popco   | corn poppy Papaver rhoeas   |
+    Scenario Outline: Searching for plants with exact/partial matches
+        When I type <plant> into the search box
+        Then the app should display <result> as first match
+        Examples:
+                | plant   | result                      |
+                | tomato  | tomato Solanum lycopersicum |
+                | tato    | potato Solanum tuberosum    |
+                | nion    | onion Allium cepa           |
+                | popco   | corn poppy Papaver rhoeas   |
 
-  Scenario: No match was found
-    When I type xyxyz into the search box
-    Then no match can be found and a text is displayed
+    Scenario: No match was found
+        When I type xyxyz into the search box
+        Then no match can be found and a text is displayed
