+++ e2e/requirements.txt
@@ -9,5 +9,3 @@ pytest-xdist
 mypy
 python-dotenv
 psycopg2-binary # only needed for development
-opencv-python
-numpy
